+++ src/components/UI/NetworkWalletButton/NetworkWalletButton.js
@@ -65,6 +65,6 @@ NetworkWalletButton.propTypes = {
   logoPath: PropTypes.string,
   network: PropTypes.string,
   status: PropTypes.string,
-  error: PropTypes.string,
+  error: PropTypes.object,
   swapFn: PropTypes.func
 };