+++ src/config/liquidity.js
@@ -6,9 +6,13 @@ export default [
     logoPath: 'liquidity/banxa.svg',
     name: 'Banxa',
     description: 'Buy crypto with credit card directly to StarkNet',
-    url: {
-      [ChainType.L1.GOERLI]: '',
-      [ChainType.L1.MAIN]: 'https://starkware.banxa.com/'
+    link: {
+      [ChainType.L1.MAIN]: {
+        url: 'https://starkware.banxa.com/',
+        qsParams: {
+          walletAddress: '{{accountL2}}'
+        }
+      }
     }
   },
   {
@@ -16,9 +20,19 @@ export default [
     logoPath: 'liquidity/layerswap.svg',
     name: 'Layerswap',
     description: 'Move crypto from Coinbase, Binance, FTX and other exchanges to StarkNet',
-    url: {
-      [ChainType.L1.GOERLI]: 'https://testnet.layerswap.io/?destNetwork=starknet_goerli',
-      [ChainType.L1.MAIN]: 'https://layerswap.io/?destNetwork=starknet_mainnet'
+    link: {
+      [ChainType.L1.GOERLI]: {
+        url: 'https://testnet.layerswap.io/',
+        qsParams: {
+          destNetwork: 'starknet_goerli'
+        }
+      },
+      [ChainType.L1.MAIN]: {
+        url: 'https://layerswap.io/',
+        qsParams: {
+          destNetwork: 'starknet_mainnet'
+        }
+      }
     }
   },
   {
@@ -26,9 +40,14 @@ export default [
     logoPath: 'liquidity/orbiter.svg',
     name: 'Orbiter',
     description: 'Cross-rollup bridge. Make deposits from other L2 rollups to StarkNet',
-    url: {
-      [ChainType.L1.GOERLI]: '',
-      [ChainType.L1.MAIN]: 'https://www.orbiter.finance/?refer=starknet&dests=starknet'
+    link: {
+      [ChainType.L1.MAIN]: {
+        url: 'https://www.orbiter.finance/',
+        qsParams: {
+          refer: 'starknet',
+          dests: 'starknet'
+        }
+      }
     }
   }
 ];