+++ src/routes/Liquidity/Liquidity.js
@@ -2,22 +2,34 @@ import React from 'react';
 
 import {ChoiceItemType, MultiChoiceMenu} from '../../components/UI';
 import {useLiquidityProviders, useLiquidityTranslation} from '../../hooks';
-import {openInNewTab, toClasses} from '../../utils';
+import {useL2Wallet} from '../../providers/WalletsProvider';
+import {openInNewTab, toClasses, buildDynamicURL} from '../../utils';
 import styles from './Liquidity.module.scss';
 
 export const Liquidity = () => {
+  const {account: accountL2} = useL2Wallet();
   const {titleTxt, descriptionTxt} = useLiquidityTranslation();
   const liquidityProviders = useLiquidityProviders();
+
+  const dynamicQsValues = {
+    accountL2
+  };
+
   const mapLiquidityProviders = () => {
-    return liquidityProviders.map(p => {
-      return {
-        ...p,
-        type: ChoiceItemType.LINK,
-        onClick: () => {
-          openInNewTab(p.url);
-        }
-      };
-    });
+    return liquidityProviders
+      .filter(p => p.link)
+      .map(p => {
+        const {link} = p;
+        const {url, qsParams} = link;
+        p.url = buildDynamicURL(url, qsParams, dynamicQsValues);
+        return {
+          ...p,
+          type: ChoiceItemType.LINK,
+          onClick: () => {
+            openInNewTab(p.url);
+          }
+        };
+      });
   };
 
   return (