+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## [1.8.0-dev.13](https://github.com/starkware-libs/starkgate-frontend/compare/v1.8.0-dev.12...v1.8.0-dev.13) (2022-08-08)
+
+
+### 🧩 Features
+
+* **2902657922:** add dynamic params to url ([#245](https://github.com/starkware-libs/starkgate-frontend/issues/245)) ([c7d614d](https://github.com/starkware-libs/starkgate-frontend/commit/c7d614d51eb04ca0721b05ea6f158cbeef1886d9))
+
+## [1.8.0-dev.12](https://github.com/starkware-libs/starkgate-frontend/compare/v1.8.0-dev.11...v1.8.0-dev.12) (2022-08-08)
+
+
+### 🔧 Fixes
+
+* **3044966975:** empty account when connecting starkgate for the 1st time ([#262](https://github.com/starkware-libs/starkgate-frontend/issues/262)) ([50f36c1](https://github.com/starkware-libs/starkgate-frontend/commit/50f36c187f5eb4e666367f6916a75b21c657fed0))
+* change error proptype from string to object ([#261](https://github.com/starkware-libs/starkgate-frontend/issues/261)) ([5b30b7a](https://github.com/starkware-libs/starkgate-frontend/commit/5b30b7a4f55cb1931b25d526b93cca28d00fa253))
+
 ## [1.8.0-dev.11](https://github.com/starkware-libs/starkgate-frontend/compare/v1.8.0-dev.10...v1.8.0-dev.11) (2022-08-08)